Change WriteShortData's function signature to enforce uint8_t on input.
Right now WriteShortData takes a uintptr_t as input and statically casts that input to a uint8_t without checking that it's a valid size for that cast.

Currently the call site for WriteShortData checks this via a DCHECK, but it would be clearer and prevent future callers from mishandling this by having WriteShortData take a uint8_t in the first place.

This was found via a CodeQL query:
https://chromium-review.googlesource.com/c/chromium/src/+/6020237

Based on a previous security bug whose root cause involved type confusion after a static cast:
https://g-issues.chromium.org/u/0/issues/351327767

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6040738
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#97973}

+++ src/codegen/reloc-info.cc
@@ -40,8 +40,8 @@ void RelocInfoWriter::WriteShortTaggedPC(uint32_t pc_delta, int tag) {
   *--pos_ = pc_delta << kTagBits | tag;
 }
 
-void RelocInfoWriter::WriteShortData(intptr_t data_delta) {
-  *--pos_ = static_cast<uint8_t>(data_delta);
+void RelocInfoWriter::WriteShortData(uint8_t data_delta) {
+  *--pos_ = data_delta;
 }
 
 void RelocInfoWriter::WriteMode(RelocInfo::Mode rmode) {
@@ -87,7 +87,7 @@ void RelocInfoWriter::Write(const RelocInfo* rinfo) {
     WriteModeAndPC(pc_delta, rmode);
     if (RelocInfo::IsDeoptReason(rmode)) {
       DCHECK_LT(rinfo->data(), 1 << kBitsPerByte);
-      WriteShortData(rinfo->data());
+      WriteShortData(static_cast<uint8_t>(rinfo->data()));
     } else if (RelocInfo::IsConstPool(rmode) ||
                RelocInfo::IsVeneerPool(rmode) || RelocInfo::IsDeoptId(rmode) ||
                RelocInfo::IsDeoptPosition(rmode) ||
